requirements.txt

Django==2.2.10
django-allauth==0.39.1
django-ckeditor==5.9.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-embed-video==1.3.3
django-filter==2.3.0
django-jet==1.0.8
django-js-asset==1.2.2
django-taggit==1.3.0
django-tinymce==3.0.2
pycodestyle==2.5.0
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-async==0.6.2
requests-oauthlib==1.2.0
rfc3986==1.4.0
six==1.15.0
sorl-thumbnail==12.6.3
sqlparse==0.2.4
stripe==2.51.0
toml==0.10.1
